alow db config generating

+++ database/db.py
@@ -17,8 +17,10 @@ ALLOWED_DESCRIPTOR_TYPES = {
 
 
 class DBConfig(object):
+    default_img_size = [256, 256]
+
     def __init__(self, path):
-        self.img_size = [256, 256]
+        self.img_size = DBConfig.default_img_size
         self.sizes = []
         self.images = []
         self.descriptor_suffix = ''
@@ -39,30 +41,56 @@ class DBConfig(object):
             if 'img_size' in data:
                 self.img_size = data['img_size']
 
+    @classmethod
+    def create_for_path(cls, path, img_suffix, sizes=None, descriptor_type=None, exclude=None):
+        if descriptor_type is None:
+            descriptor_type = '.lbp'
+        if descriptor_type not in ALLOWED_DESCRIPTOR_TYPES:
+            raise Exception('No such descriptor')
+
+        if sizes is None:
+            sizes = [[16, 16]]
+
+        images = []
+        for fname in os.listdir(path):
+            if fname.endswith(img_suffix):
+                images.append(fname)
+
+        with open(os.path.join(path, 'db.txt'), 'wb') as f:
+            data = {"images": images,
+                    "sizes": sizes,
+                    "descriptor_suffix": descriptor_type}
+            f.write(json.dumps(data))
+
 
 ELEMENT_SIZE = 4
 
 
 class DescriptorsDB(object):
-    def __init__(self):
+    def __init__(self, db_location=DATABASE_LOCATION):
         self.images = dict()
         self.cfg = None
-        db_path = os.path.join(DATABASE_LOCATION, 'db.txt')
-        self.cfg = DBConfig(db_path)
+        self.db_location = db_location
+        self.db_path = os.path.join(db_location, 'db.txt')
+        self.cfg = DBConfig(self.db_path)
         self.load_metainfo()
 
     def _get_descriptor_location(self, image_name):
-        return os.path.join(DATABASE_LOCATION, image_name) + self.cfg.descriptor_suffix
+        return os.path.join(self.db_location, image_name) + self.cfg.descriptor_suffix
 
     def _get_descriptor(self):
         return ALLOWED_DESCRIPTOR_TYPES[self.cfg.descriptor_suffix]()
 
     def load_metainfo(self):
-        db_path = os.path.join(DATABASE_LOCATION, 'db.txt')
+        db_path = os.path.join(self.db_location, 'db.txt')
         self.cfg = DBConfig(db_path)
         for image_name in self.cfg.images:
             image_location = self._get_descriptor_location(image_name)
             self.images.update({image_location: os.path.exists(image_location)})
+
+    def make_all_descriptors(self, force=False):
+        for img in self.cfg.images:
+            self.make_descriptors(img)
 
     def calculate_descriptors(self, img_path):
         descriptor = self._get_descriptor()
@@ -81,8 +109,10 @@ class DescriptorsDB(object):
         img_part = crop_image(w * i, h * j, w * (i + 1), h * (j + 1), img)
         return descriptor.calculate_descriptor(img_part)
 
-    def make_descriptors(self, img_id):
-        img_path = DATABASE_LOCATION + img_id
+    def make_descriptors(self, img_id, force=False):
+        img_path = os.path.join(self.db_location, img_id)
+        if img_path in self.images and self.images[img_path] and not force:
+            return
         _descriptor = self._get_descriptor()
         pack_template = '{}f'.format(_descriptor.size())
         with open(img_path + self.cfg.descriptor_suffix, 'wb') as descrfile:
